assignment4/requirements.txt
argparse==1.4.0
Cython==0.29.24
Pillow==8.3.2
cycler==0.10.0
kiwisolver==1.3.2
llvmlite==0.37.0
matplotlib==3.4.3
numba==0.54.1
numpy==1.20.3
opencv_python==4.5.3.56
pip==21.2.4
pyparsing==2.4.7
python_dateutil==2.8.2
setuptools==58.0.4
six==1.16.0
wheel==0.37.0
attrs-21.2.0
iniconfig-1.1.1
packaging-21.0
pluggy-1.0.0
py-1.10.0
pytest-6.2.5
toml-0.10.2
